+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - `isFunction` function.
 - `Typeable`, `Sizeable`, `Semogroupoid`, `Semigroup` and `Compressable` interfaces.
 - `apply` method to `Stream`.
+- `concat` method to `Tuple`.
 
 ### Changed